refactor: Split SQLLogicTest adapter into focused modules
Split the monolithic 983-line db_adapter.rs into smaller, focused modules:

- **pool.rs** (52 lines): Thread-local database pooling logic
- **timing.rs** (185 lines): Statement timing and performance instrumentation
- **cache.rs** (74 lines): Query result caching and invalidation
- **batching.rs** (71 lines): Automatic INSERT batching for test data loading
- **executor.rs** (550 lines): Core SQL execution and statement dispatching
- **adapter.rs** (254 lines): Main VibeSqlDB implementation (thin glue code)
- **mod.rs** (18 lines): Module declarations and re-exports

Benefits:
- Separation of concerns: Each module has one clear responsibility
- Easier testing: Can test caching, timing, batching independently
- Improved maintainability: Changes isolated to specific modules
- Better code organization: Developers can find functionality quickly
- Reduced file size: Each module 50-550 lines vs 983 monolithic

All functionality preserved with no API changes.
